+++ test/curator/hooks_test.exs
@@ -1,4 +1,0 @@
-defmodule Curator.HooksTest do
-  use ExUnit.Case, async: true
-  doctest Curator.Hooks
-end