+++ test/curator/plug/ensure_resource_and_session_test.exs
@@ -1,4 +1,0 @@
-defmodule Curator.Plug.EnsureResourceAndSessionTest do
-  use ExUnit.Case, async: true
-  doctest Curator.Plug.EnsureResourceAndSession
-end