+++ test/curator/plug/ensure_resource_or_no_session_test.exs
@@ -1,4 +1,0 @@
-defmodule Curator.Plug.EnsureResourceOrNoSessionTest do
-  use ExUnit.Case, async: true
-  doctest Curator.Plug.EnsureResourceOrNoSession
-end