+++ test/curator/plug_helper_test.exs
@@ -1,4 +1,0 @@
-defmodule Curator.PlugHelperTest do
-  use ExUnit.Case, async: true
-  doctest Curator.PlugHelper
-end