+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v 0.2.0
-Rewrite
+Rewrite - currently only supports an Ueberauth workflow (and timeouts). v0.3.0 will add support back for database authentication.
 
 * Generators match Phoenix 1.3 structure
 * Requires Guardian 1.0